+++ Changelog.md
@@ -10,8 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Allow empty optional services list.
+- Allow passing empty option value via universal options like `--with-web-root=""`.
 - Make upgrade.sh more compatible with the latest Git versions.
-
 
 
 ## [3.1.2] - 2022-09-08